--- opencv_util/setup.py
@@ -1,11 +1,0 @@
-#!/usr/bin/env python
-from distutils.core import setup
-from catkin_pkg.python_setup import generate_distutils_setup
-
-d = generate_distutils_setup(
-    packages=['opencv_util'],
-    package_dir={'': 'src'},
-    requires=['genmsg', 'genpy', 'roslib', 'rospkg']
-)
-
-setup(**d)
