+++ swri_opencv_util/include/swri_opencv_util/show.h
@@ -32,7 +32,7 @@
 
 #include <opencv2/core/core.hpp>
 
-namespace opencv_util
+namespace swri_opencv_util
 {
   void ShowScaled(
       const std::string& name,
